requirements.txt
agno>=2.4.7
anthropic>=0.72.0
google-genai>=1.47.0
google-generativeai>=0.8.5
openai>=2.6.1
mcp>=1.21.0
rich>=14.2.0
python-dotenv>=1.2.1
sqlalchemy
tavily-python
httpx
# Telegram API
telethon==1.34.0

playwright

# Optional: multi-provider model routing
# litellm